refactor(web): scope shadcn/theme files by domain and document the alias
- Move cn() to web/src/lib/shadcn/utils.ts (+ components.json alias) and the
  color mapper to web/src/theme/http-color-mapper.ts.
- Extract the shadcn theme tokens to web/src/lib/shadcn/styles.css; index.css
  now just imports Tailwind + those tokens.
- Comment the @ alias in vite/vitest configs and the "*.css" module declaration.
