+++ web/src/components/ui/label.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { Label as LabelPrimitive } from "radix-ui"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/shadcn/utils"
 
 function Label({
   className,
